requirements.txt
streamlit
SpeechRecognition
google-generativeai
python-dotenv
audio-recorder-streamlit